Sip Dependencies in the SIP.App project so that it can be used without all the Web and Database references and can target netstandard2.0.
