+++ sipsorcery-core/SIPSorcery-Core.sln
@@ -21,7 +21,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SIPSorcery.Sys", "SIPSorcer
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SIPSorcery.SIP.Core", "SIPSorcery.SIP.Core\SIPSorcery.SIP.Core.csproj", "{01D01BC6-231F-4FD6-BADF-51C509AB495F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SIPSorcery.SIP.App", "SIPSorcery.SIP.App\SIPSorcery.SIP.App.csproj", "{339648E6-3B61-4A27-B4A2-38A8216B1D7E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SIPSorcery.SIP.App", "SIPSorcery.SIP.App\SIPSorcery.SIP.App.csproj", "{339648E6-3B61-4A27-B4A2-38A8216B1D7E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SIPSorcery.Net", "SIPSorcery.Net\SIPSorcery.Net.csproj", "{55898587-8456-40D4-BB88-C9FC8BD6BA9C}"
 EndProject
