+++ sipsorcery-core/SIPSorcery.SIP.App/SIPUserAgents/SIPTransferServerUserAgent.cs
@@ -56,7 +56,7 @@ namespace SIPSorcery.SIP.App
         {
             get { throw new NotImplementedException("SIPTransferServerUserAgent SIPDialogue"); }
         }
-        public SIPAccount SIPAccount
+        public ISIPAccount SIPAccount
         {
             get
             {
